+++ tk/src/circleobject.cpp
@@ -14,7 +14,7 @@ CircleObject::~CircleObject() { };
 void CircleObject::drawWorld() {
 }
 
-void CircleObject::drawScreen(bool drawCenterPoint, float cellsPerMeter, int pixelsPerCell, int doriScreenX, int doriScreenY) {
+void CircleObject::drawScreen(bool drawCenterPoint, float cellsPerMeter, float pixelsPerCell, int doriScreenX, int doriScreenY) {
     // # cells = (cells / meter) * (worldOffset in meters)
     // # cells * pixelsPerCell = position in pixels
     float screenOffsetX = (cellsPerMeter * worldOffsetX) * pixelsPerCell;
